requirements.txt
elasticsearch[async]==7.10.1
lxml==4.6.1
PyYAML==5.3.1
requests==2.25.1
tqdm==4.53.0
